+++ lib/glimmer/libui/control_proxy/menu_item_proxy.rb
@@ -41,3 +41,5 @@ module Glimmer
     end
   end
 end
+
+Dir[File.expand_path('./menu_item_proxy/*.rb', __dir__)].each {|f| require f}