+++ lib/glimmer/libui/control_proxy/menu_item_proxy/quit_menu_item_proxy.rb
@@ -24,14 +24,41 @@ require 'glimmer/libui/control_proxy/menu_item_proxy'
 module Glimmer
   module LibUI
     class ControlProxy
-      # Proxy for LibUI check menu item object
-      #
-      # Follows the Proxy Design Pattern
-      class CheckMenuItemProxy < MenuItemProxy
-        private
+      class MenuItemProxy < ControlProxy
+        # Proxy for LibUI quit menu item object
+        #
+        # Follows the Proxy Design Pattern
+        class QuitMenuItemProxy < MenuItemProxy
+          def can_handle_listener?(listener_name)
+            listener_name == 'on_clicked' || super
+          end
         
-        def build_control
-          @libui = @parent_proxy.append_check_item(*@args)
+          def handle_listener(listener_name, &listener)
+            if listener_name == 'on_clicked'
+              @default_behavior_listener = Proc.new do
+                return_value = listener.call(self)
+                if return_value.is_a?(Numeric)
+                  return_value
+                else
+                  destroy
+                  ::LibUI.quit
+                  0
+                end
+              end
+              ::LibUI.on_should_quit(&@default_behavior_listener)
+            end
+          end
+        
+          private
+          
+          def build_control
+            @libui = @parent_proxy.append_quit_item(*@args)
+            handle_listener('on_clicked') do
+              ControlProxy.main_window_proxy&.destroy
+              ::LibUI.quit
+              0
+            end
+          end
         end
       end
     end